backend/requirements.txt
# PharmaPricing Backend - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
PyJWT==2.8.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development
black==23.12.0
flake8==6.1.0
pylint==3.0.3
isort==5.13.2

# Production
gunicorn==21.2.0
